cleanup - convert hook to a util function
